+++ NEWS.md
@@ -1,6 +1,12 @@
 # Release Notes
 
 
+## v0.3.0
+
+- Import/export using Tables.jl interface.
+- Added `DataKnot(::Pair...)` constructor.
+
+
 ## v0.2.0
 
 - Added `Unique` and `Group` combinators.